Login/requirements.txt
bcrypt~=4.2.1
Flask~=3.1.0
Flask-Login~=0.6.3
python-dotenv~=1.0.1